Added more to the Red map.
Added color_tiles.png to TileMap assets.
Added color_tiles.tsx to TileMap assets.
This was needed so that editing the maps was easier for the paths of the Tilesets.
Added Red.tmx
This was needed so that I could eidt and keep track of the Red map.
Added Tutorial.tmx
This was needed because the Tutorial was being saved elsewhere in the path, fixed its location.
Removed TutorialMap.json
Not required anymore.
Moved path for Red.tmx
Moved Path for Tutorial.tmx
Removed outdated Tutorial.tmx
Edited Red.js
Added the red upgrade, which allows players to shoot bullets.
Added the go home button which allows players to go back to the Tutorial.