Edited the Maps
Edited Blue.tmx - Added the framework for the leve.
Edited Blue.json - Exported the .tmx as the .json to update fhe map.
Edited Red.tmx - Changed one of the platforms.
Edited Red.json - Exported .tmx of Red to update Red level.
Edited Tutorial.tmx - Added the positioning for all of the colors, and King Color.
Edited Tutorial.json - Exported the .tmx to update the map.
Edited Yellow.tmx - Added the framework for Yellow.
Edited Yellow.json - Exported the .tmx to update the map.
Edited Main.js - Removed the player starting with all of the powers.
Edited Blue.js - to implement the edits of the map, as well as re-positioning things.
Edited Red.js - to implement the edits of the map, as well as re-positioning things.
Edited Yellow.js - to implement the edits of the map, as well as re-positioning things.
Edited Tutorial.js - to implement and position all of the color portals to where they should be.