BIG PATCH OF CHANGES
Added Sad1.mp3 - new music, designed to be the first sad music that the player interacts with. The music will change as they get more colors.
Added Bucky2.png - This was intended to be used for keeping track of Bucky's direction for the time being. It didn't quite work out.
Added window.json - This is the JSON file for the window atlas that I added to keep track of the raining window asset and animation.
Added window_01/02/03.png - This was to individualize the frames of the window animation, because the atlas wasn't working at first.
Edited color_tiles.png - Changed it to the current tiles of the game, including objects and shelves.
Edited Tutorial.json - Edited the level to include the new tiles.
Edited Tutorial.tmx - Edited the level to include the new tiles.
Edited color_tiles.tsx - Edited the tiles that will be used in the game.
Edited Player.js - Changed the scale of Bucky to be more easily visible, and very nice to look at.
Edited Load.js - Changed out the music that would be loaded, added Bucky2, and added the window.
Edited Tutorial.js - Changed the level to fit the new map.