Edited Maps
Edited Blue.json - Adjusted the levels to fit the player more easily.
Edited Blue.tmx - See Blue.json.
Edited Blue.js - Changed the groups and added the collisions between the jumping enemies and the level. Also added the collision between the jumping enemies and the player.
Edited BossMap.js - Implemented the BossMap so that it actually loads KingColor.json.
Edited Yellow.js - Made it so that the health should be visible (?) Eric should take a look at it so that it works.