Added Post Bossfight content
Edited Tutorial.json - Added "Objects" in the object layer of the map so that we can position the NPCs with ease.
Edited Tutorial.tmx - See Tutorial.json.
Edited Tutorial.js - Used the bossDefeared boolean to create the scenario for the world after the bossfight. Categorized the NPCs so that they can first spawn in their proper positions, and then spawn at the end of the map after the bossfight. This is done so that the player can both see the credits of the game, and be congratulated by the NPCs.