Added the credits, and fixed the yellow level
Added Daniel.png - The picture to represent Daniel in the credits of the game.
Added Dylan.png - The picture to represent Dylan in the credits of the game.
Added EricAndreae.png - The picture to represent Eric Andreae in the credits of the game.
Added EricHu.png - The picture to represent Eric Hu in the credits of the game.
Edited Yellow.json - Balanced out yellow a bit, and made it so that there are 3 enemies in level 2.
Edited Yellow.tmx - See Yellow.json.
Edited Load.js - Added the credit images to be loaded.
Edited Tutorial.js - Setup the credit photos and the credit text in the final level of the game.
Edited Yellow.js - Made it so that there are 3 enemies in the game, also fixed the yellow blobs spawning too high off the ground.